refactor : moved item files into a item directory
